Use timestamp to guarantee unique IDs in create forest
The create forest command is generated with `"DEPENDENTJOB%i" % i` where
`i` is the value of the iterator. If the command is fun multiple times,
it uses the same names over and over. `--create` appends
`datetime.datetime.utcnow()`, which has resolution to 1/100,000th of a
second and is granular enough to produce unique values between
iterations in the loop.

This fixes --createforest to produce unique job names.